pyproject.toml
[tool]
[tool.poetry]
name = "modularyze"
version = "0.1.0"
homepage = "https://github.com/jungerm2/modularyze"
description = "Top-level package for modularyze."
authors = ["Sacha Jungerman <jungerm2@illinois.edu>"]
readme = "README.rst"
license =  "GPL-3.0-only"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "modularyze" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = "^3.6"
"ruamel.yaml" = "^0.16.12"
Jinja2 = "^2.11.2"

[tool.poetry.dev-dependencies]
bumpversion = "*"
coverage = "*"
flake8 = "*"
invoke = "*"
pylint = "*"
pytest = "*"
sphinx = "*"
tox = "*"
testfixtures = "^6.15.0"
sphinx-rtd-theme = "^0.5.0"
black = {version = "19.3b0", allow-prereleases = true}

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
